Quick Filters, pages minor modifications

+++ Models/ProductVM.cs
@@ -7,9 +7,9 @@ namespace DentalApp.Models
         public string? Name { get; set; }
         public string? Description { get; set; }
         public string? ProductCode { get; set; }
-        public double Amount { get; set; } = 0.00;
-        public double MinPrice { get; set; } = 0.00;
-        public double MaxPrice { get; set; } = 0.00;
+        public decimal Amount { get; set; } = 0.00m;
+        public decimal MinPrice { get; set; } = 0.00m;
+        public decimal MaxPrice { get; set; } = 0.00m;
         public ProductType ProductType { get; set; }
     }
 }
